project-2/requirements/base.txt
Django==3.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-imagekit==4.0.2
Pillow==7.2.0
psycopg2==2.8.5